+++ starters/gatsby-starter-catalyst-bery/gatsby-config.js
@@ -52,7 +52,6 @@ module.exports = {
         // Core Theme
         // contentPath: `content/pages`,
         // assetPath: `content/assets`,
-        // useKatex: false,
         // remarkImagesWidth: 1440,
         // imageQuality: 50,
         // useAlertBanner: false,