+++ starters/gatsby-starter-catalyst-blog/gatsby-config.js
@@ -58,7 +58,6 @@ module.exports = {
         //Default options are:
         // contentPath: `content/pages`,
         // assetPath: `content/assets`,
-        // useKatex: false,
         // remarkImagesWidth: 1440,
         // imageQuality: 50,
         // useAlertBanner: false,