+++ themes/gatsby-theme-catalyst-bery/src/components/nav.js
@@ -15,6 +15,7 @@ const Nav = () => {
         display: ["none", "block", null, null, null],
         gridColumn: "1 / -1",
         mt: 4,
+        fontSize: 2,
         variant: "variants.nav",
       }}
       aria-label="Primary menu"