+++ themes/gatsby-theme-catalyst-blog/gatsby-config.js
@@ -43,7 +43,7 @@ module.exports = (themeOptions) => {
                   allCatalystPost(
                     sort: { fields: [date, title], order: DESC }
                     limit: 1000
-                    filter: { draft: { eq: false } }
+                    filter: {draft: {ne: true}, published: {eq: true}}
                   ) {
                     nodes {
                       id