+++ www/src/components/header/branding.js
@@ -34,8 +34,6 @@ const SiteBranding = () => {
             variant: "variants.siteLogo",
           }}
           alt={title}
-          objectFit="contain"
-          loading="eager"
           imgStyle={{ objectFit: "contain" }}
         />
       </Link>
@@ -48,9 +46,14 @@ const SiteBranding = () => {
           display: ["none", null, "block", null, null],
         }}
       >
-        <Themed.h1
-          as="span"
+        <span
           sx={{
+            fontWeight: "500",
+            textTransform: "uppercase",
+            fontSize: "1.5rem !important",
+            textAlign: "center",
+            mt: 3,
+            letterSpacing: "wide",
             color: isNavOpen ? "header.textOpen" : "header.text",
             textDecoration: "none",
             m: 0,
@@ -58,7 +61,7 @@ const SiteBranding = () => {
           }}
         >
           {title}
-        </Themed.h1>
+        </span>
       </Link>
     </div>
   )