+++ starters/gatsby-starter-catalyst-core/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## 6.0.0-beta.10
 
-## [3.1.3](https://github.com/ehowey/gatsby-starter-catalyst-core/compare/gatsby-starter-catalyst-core@3.1.2...gatsby-starter-catalyst-core@3.1.3) (2021-02-18)
+### Patch Changes
 
-**Note:** Version bump only for package gatsby-starter-catalyst-core
+- c76af932: Adds support for `published: true` boolean to MDX frontmatter. This flag is now required for blog posts to be published. Prevents accidentally publishing blog posts.
+- Updated dependencies [c76af932]
+  - gatsby-theme-catalyst-core@6.0.0-beta.10
+
+## 6.0.0-beta.8
+
+### Patch Changes
+
+- 1d2e3b90: Update Gatsby to v3.4 and Theme-UI to v0.7.5. Updated additional packages as necessary.
+- Updated dependencies [1d2e3b90]
+  - gatsby-theme-catalyst-core@6.0.0-beta.8
+
+## 6.0.0-beta.7
+
+### Patch Changes
+
+- d4969468: Remove loading=eager prop as this is causing issues with blurry images using gatsby-plugin-image
+- Updated dependencies [d4969468]
+  - gatsby-theme-catalyst-core@6.0.0-beta.7