+++ themes/gatsby-theme-catalyst-lithium/CHANGELOG.md
@@ -1,8 +1,44 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## 6.0.0-beta.10
 
-## [3.1.3](https://github.com/ehowey/gatsby-theme-catalyst/compare/gatsby-theme-catalyst-lithium@3.1.2...gatsby-theme-catalyst-lithium@3.1.3) (2021-02-18)
+### Patch Changes
 
-**Note:** Version bump only for package gatsby-theme-catalyst-lithium
+- c76af932: Adds support for `published: true` boolean to MDX frontmatter. This flag is now required for blog posts to be published. Prevents accidentally publishing blog posts.
+- Updated dependencies [c76af932]
+  - gatsby-theme-catalyst-blog@6.0.0-beta.10
+  - gatsby-theme-catalyst-core@6.0.0-beta.10
+  - gatsby-theme-catalyst-footer@6.0.0-beta.10
+  - gatsby-theme-catalyst-header-top@6.0.0-beta.10
+
+## 6.0.0-beta.9
+
+### Patch Changes
+
+- 3f8071db: Updated to Theme-UI v0.8.4 and fixes all necessary styles that were affected by moving the root styles from the body element to the root element. This may affect your text sizing if you are using this in an existing site.
+- Updated dependencies [3f8071db]
+  - gatsby-theme-catalyst-core@6.0.0-beta.9
+  - gatsby-theme-catalyst-footer@6.0.0-beta.9
+  - gatsby-theme-catalyst-header-top@6.0.0-beta.9
+
+## 6.0.0-beta.8
+
+### Patch Changes
+
+- 1d2e3b90: Update Gatsby to v3.4 and Theme-UI to v0.7.5. Updated additional packages as necessary.
+- Updated dependencies [1d2e3b90]
+  - gatsby-theme-catalyst-blog@6.0.0-beta.8
+  - gatsby-theme-catalyst-core@6.0.0-beta.8
+  - gatsby-theme-catalyst-footer@6.0.0-beta.8
+  - gatsby-theme-catalyst-header-top@6.0.0-beta.8
+
+## 6.0.0-beta.7
+
+### Patch Changes
+
+- d4969468: Remove loading=eager prop as this is causing issues with blurry images using gatsby-plugin-image
+- Updated dependencies [d4969468]
+  - gatsby-theme-catalyst-blog@6.0.0-beta.7
+  - gatsby-theme-catalyst-core@6.0.0-beta.7
+  - gatsby-theme-catalyst-footer@6.0.0-beta.7
+  - gatsby-theme-catalyst-header-top@6.0.0-beta.7