+++ themes/gatsby-theme-catalyst-sanity/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## 6.0.0-beta.10
 
-## [4.0.8](https://github.com/ehowey/gatsby-theme-catalyst/compare/gatsby-theme-catalyst-sanity@4.0.7...gatsby-theme-catalyst-sanity@4.0.8) (2021-02-18)
+### Patch Changes
 
-**Note:** Version bump only for package gatsby-theme-catalyst-sanity
+- c76af932: Adds support for `published: true` boolean to MDX frontmatter. This flag is now required for blog posts to be published. Prevents accidentally publishing blog posts.
+
+## 6.0.0-beta.8
+
+### Patch Changes
+
+- 1d2e3b90: Update Gatsby to v3.4 and Theme-UI to v0.7.5. Updated additional packages as necessary.
+
+## 6.0.0-beta.7
+
+### Patch Changes
+
+- d4969468: Remove loading=eager prop as this is causing issues with blurry images using gatsby-plugin-image