+++ themes/gatsby-theme-catalyst-stripe/CHANGELOG.md
@@ -1,8 +1,34 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## 0.1.0-beta.10
 
-## 0.0.7 (2021-02-18)
+### Patch Changes
 
-**Note:** Version bump only for package gatsby-theme-catalyst-stripe
+- c76af932: Adds support for `published: true` boolean to MDX frontmatter. This flag is now required for blog posts to be published. Prevents accidentally publishing blog posts.
+- Updated dependencies [c76af932]
+  - gatsby-theme-catalyst-core@6.0.0-beta.10
+  - gatsby-theme-catalyst-footer@6.0.0-beta.10
+  - gatsby-theme-catalyst-header-top@6.0.0-beta.10
+  - gatsby-theme-catalyst-sanity@6.0.0-beta.10
+
+## 0.1.0-beta.8
+
+### Patch Changes
+
+- 1d2e3b90: Update Gatsby to v3.4 and Theme-UI to v0.7.5. Updated additional packages as necessary.
+- Updated dependencies [1d2e3b90]
+  - gatsby-theme-catalyst-core@6.0.0-beta.8
+  - gatsby-theme-catalyst-footer@6.0.0-beta.8
+  - gatsby-theme-catalyst-header-top@6.0.0-beta.8
+  - gatsby-theme-catalyst-sanity@6.0.0-beta.8
+
+## 0.1.0-beta.7
+
+### Patch Changes
+
+- d4969468: Remove loading=eager prop as this is causing issues with blurry images using gatsby-plugin-image
+- Updated dependencies [d4969468]
+  - gatsby-theme-catalyst-core@6.0.0-beta.7
+  - gatsby-theme-catalyst-footer@6.0.0-beta.7
+  - gatsby-theme-catalyst-header-top@6.0.0-beta.7
+  - gatsby-theme-catalyst-sanity@6.0.0-beta.7